+++ src/main/java/org/magic/api/interfaces/MTGExternalShop.java
@@ -14,12 +14,12 @@ public interface MTGExternalShop extends MTGPlugin {
 	
 	
 	public List<MTGProduct> listProducts(String name) throws IOException;
-	public int createProduct(MTGProduct t,Category c) throws IOException;
-	public int createProduct(MTGExternalShop input, MTGProduct t, String lang,Category c)throws IOException;
+	public Long createProduct(MTGProduct t,Category c) throws IOException;
+	public Long createProduct(MTGExternalShop input, MTGProduct t, String lang,Category c)throws IOException;
 	
 	
 	public List<MTGStockItem> listStock(String search) throws IOException;
-	public MTGStockItem getStockById(EnumItems typeStock,Integer id)throws IOException;
+	public MTGStockItem getStockById(EnumItems typeStock,Long id)throws IOException;
 	public void saveOrUpdateStock(MTGStockItem stock,boolean allShop)throws IOException;
 	public void saveOrUpdateStock(List<MTGStockItem> stocks,boolean allShop)throws IOException;
 	
@@ -40,7 +40,7 @@ public interface MTGExternalShop extends MTGPlugin {
 	public void deleteTransaction(List<Transaction> t) throws IOException;
 	public Transaction getTransactionById(int parseInt) throws IOException;
 	
-	public void updateConversion(String name, String destName, String language, Integer idProduct, int idDestProduct) throws IOException;
+	public void updateConversion(String name, String destName, String language, Long idProduct, Long idDestProduct) throws IOException;
 	public List<Transaction> listTransactions(Contact c) throws IOException;
 	public boolean enableContact(String token)throws IOException;
